+++ tests/symbols.clj
@@ -1,0 +1,4 @@
+'clojure.core/map
+
+'example/foo
+