+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.17.0 - 2023-02-25
+- Goto Symbol definition
+- Find Symbol usages
+- Show Symbol documentation
+
+## 0.16.1 - 2022-12-22
+- Fix None Window in `project_path`
+- Fix **Trace Usages** Window flags
+
 ## 0.16.0 - 2022-11-25
 - Support Clojure Sublimed